Added functionality for editing and deleting tasks in task list

+++ src/main/java/com/todolist/service/TaskService.java
@@ -4,12 +4,15 @@ import com.todolist.dto.TaskDTO;
 import com.todolist.dto.mapper.TaskMapper;
 import com.todolist.model.Task;
 import com.todolist.model.TaskList;
+import com.todolist.model.User;
 import com.todolist.repository.TaskRepository;
 import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 
 @Service
@@ -19,12 +22,15 @@ public class TaskService {
     private final TaskListService taskListService;
 
     @Autowired
-    public TaskService(TaskRepository taskRepository, TaskListService taskListService) {
+    public TaskService(TaskRepository taskRepository, @Lazy TaskListService taskListService) {
         this.taskRepository = taskRepository;
         this.taskListService = taskListService;
     }
 
-    public TaskDTO createTask(TaskDTO taskDTO, Long taskListId) throws NotFoundException {
+    public TaskDTO createTask(TaskDTO taskDTO, Long taskListId, User user) throws NotFoundException {
+        if (!taskListService.userContainsTaskListId(user, taskListId)) {
+            throw new NotFoundException("TaskList with id " + taskListId + " not found for the user");
+        }
         TaskList taskList = taskListService.findTasklist(taskListId);
         Task task = TaskMapper.mapToTask(taskDTO);
         task.setDateTime(LocalDateTime.now());
@@ -33,8 +39,11 @@ public class TaskService {
         return taskDTO;
     }
 
-    public TaskDTO editTask(TaskDTO taskDTO, Long idTask) throws Exception {
-        Task task = taskRepository.findByIdTask(idTask).orElseThrow(() -> new NotFoundException("Task not found"));;
+    public TaskDTO editTask(TaskDTO taskDTO, Long idTask, User user, Long taskListId) throws Exception {
+        Task task = taskRepository.findByIdTask(idTask).orElseThrow(() -> new NotFoundException("Task not found"));
+        if (!taskListService.userContainsTaskListId(user, taskListId)) {
+            throw new NotFoundException("TaskList with id " + taskListId + " not found for the user");
+        }
         task.setPriorityTask(taskDTO.getPriorityTask());
         task.setDescription(taskDTO.getDescription());
         task.setDateTime(LocalDateTime.now());
@@ -42,9 +51,33 @@ public class TaskService {
         return TaskMapper.mapToDTO(task);
     }
 
-    public void deleteTask(Long idTask) throws Exception {
+    public void deleteTask(Long idTask, User user, Long idTaskList) throws Exception {
         Task task = taskRepository.findByIdTask(idTask).orElseThrow(() -> new NotFoundException("Task not found"));
+        if (!taskListService.userContainsTaskListId(user, idTaskList)) {
+            throw new NotFoundException("TaskList with id " + idTaskList + " not found for the user");
+        }
         taskRepository.delete(task);
     }
 
+    public void deleteTasksByTaskList(Long idTaskList, User user) throws Exception {
+        if (!taskListService.userContainsTaskListId(user, idTaskList)) {
+            throw new NotFoundException("TaskList with id " + idTaskList + " not found for the user");
+        }
+
+        List<Task> tasksToDelete = taskRepository.findByTaskListId(idTaskList);
+
+        if (tasksToDelete.isEmpty()) {
+            throw new NotFoundException("No tasks found for TaskList with id " + idTaskList);
+        }
+
+        taskRepository.deleteAll(tasksToDelete);
+    }
+
+    public TaskDTO getTask(Long idTask, Long idTaskList, User user) throws NotFoundException {
+        if (!taskListService.userContainsTaskListId(user, idTaskList)) {
+            throw new NotFoundException("TaskList with id " + idTaskList + " not found for the user");
+        }
+        Task task = taskRepository.findByIdTask(idTask).orElseThrow(() -> new NotFoundException("Task not found"));
+        return TaskMapper.mapToDTO(task);
+    }
 }
